Form vide + optimistation and break

+++ includes/deleteMerchant.php
@@ -5,14 +5,11 @@ if (!class_exists('Connection')) {
 }
 $response = [];
 
-    switch ($_SESSION["profil"]) {
-        case 'PO':
+switch ($_SESSION["profil"]) {
+    case 'PO':
+        if (!empty($_POST['login']) && !empty($_POST['companyName'])) {
             $delete = "INSERT INTO TRAN_REQUEST_PO (login, companyName, comment, type) VALUES (:login, :companyName, :comment, :type)";
-            if (!isset(($_POST['comment']))) {
-                $comment = "";
-            } else {
-                $comment = $_POST['comment'];
-            }
+            $comment = !empty($_POST['comment']) ? $_POST['comment'] : "";
             $conditions = array(array(':login', $_POST['login']), array(':companyName', $_POST['companyName']), array(':comment', $comment), array(':type', '1'));
             $queryDelete = $db->query($delete, $conditions);
 
@@ -20,7 +17,9 @@ $response = [];
             echo json_encode($response);
 
             break;
-        case 'Admin':
+        }
+    case 'Admin':
+        if (!empty($_POST['login'])) {
             $suppMerchant = "DELETE FROM TRAN_USERS WHERE login = :login";
             $conditions = array(array(':login', $_POST['login']));
             $query3 = $db->query($suppMerchant, $conditions);
@@ -29,5 +28,5 @@ $response = [];
             echo json_encode($response);
 
             break;
-    }
-?>
+        }
+}
